Works kinda

+++ srcs/requirements/wordpress/conf/script.sh
@@ -1,14 +1,47 @@
 #!/bin/bash
+
+set -e 
+
 cd /var/www/html
 
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
-chmod +x wp-cli.phar
-curl https://wordpress.org/latest.tar.gz -P /var/www
-tar -xzvf /var/www/latest.tar.gz -C /var/www/html --strip-components=1
-rm /var/www/latest.tar.gz
+# Ensure the PHP run directory exists
+if [ ! -d /run/php ]; then
+    mkdir /run/php
+fi
 
-wp core download --allow-root
-wp config create --dbname=wordpress --dbuser=wpuser --dbpass=password --dbhost=mariadb --allow-root
-wp core install --url=localhost --title=inception --admin_user=admin --admin_password=admin --admin_email=admin@admin.com --allow-root
+# Ensure wp-cli is installed
+if [ ! -f /usr/local/bin/wp ]; then
+    curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+    chmod +x wp-cli.phar
+    mv wp-cli.phar /usr/local/bin/wp
+fi
 
+# Download and extract WordPress if it's not already installed
+if [ ! -d wp-admin ]; then
+    wget -q https://wordpress.org/latest.tar.gz -P .
+    tar -xzf latest.tar.gz --strip-components=1
+    rm -rf latest.tar.gz
+fi
+
+# Create the wp-config.php file if it doesn't exist
+if [ ! -f wp-config.php ]; then
+	# wp core download --allow-root
+    wp config create --dbname=$WP_DB \
+                     --dbuser=$WP_USER \
+                     --dbpass=$WP_PASSWORD \
+                     --dbhost=$HOSTNAME \
+                     --path='/var/www/html' \
+                     --allow-root
+
+    wp core install --url=$URL \
+                    --title=inception \
+                    --admin_user=$WP_ADMIN_USER \
+                    --admin_password=$WP_ADMIN_PASSWORD \
+                    --admin_email=$WP_ADMIN_EMAIL \
+                    --allow-root
+
+    wp user create $WP_USER $WP_EMAIL --role=author --user_pass=$WP_PASSWORD --allow-root
+fi
+
+# Start PHP-FPM
 php-fpm7.4 -F
